pyproject.toml
# pyproject.toml (version finale et correcte)

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pycomfy"
version = "0.4.0" # Version incrémentée pour les nouvelles fonctionnalités
authors = [
  { name="JusteLeo" },
]
description = "A simple and robust Python client library for the ComfyUI API."
readme = "README.md"
license = "MIT"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "requests",
    "websocket-client",
    "Pillow",
    "huggingface-hub" # Dépendance ajoutée
]

[project.urls]
Homepage = "https://github.com/votre_nom/pycomfy"
Issues = "https://github.com/votre_nom/pycomfy/issues"

[tool.setuptools]
include-package-data = true